# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GenAI Email & Report Drafting API
  version: 1.0.0
  description: Flask REST API for AI-powered email and report drafting.
servers:
  - url: /
    description: Same origin as the Swagger UI host
paths:
...
  /api/history:
    get:
      summary: Get document history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: doc_type
          schema:
            type: string
            enum: [email, report]
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/history/{document_id}:
    get:
      summary: Get a specific document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  document:
                    $ref: "#/components/schemas/Document"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: string
...
    Document:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        doc_type:
          type: string
          enum: [email, report]
        title:
          type: string
          nullable: true
        prompt_input:
          type: string
        content:
          type: string
        tone:
          type: string
        structure:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    DocumentSummary:
      type: object
      properties:
        id:
          type: integer
        doc_type:
          type: string
          enum: [email, report]
        title:
          type: string
          nullable: true
        tone:
          type: string
        structure:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        content_preview:
          type: string
...
    HistoryResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
...